+++ clint/__init__.py
@@ -11,7 +11,14 @@ This module sets up the main interface for all of clint.
 
 from __future__ import absolute_import
 
-from . import arguments
+try:
+    from collections import OrderedDict
+except ImportError:
+    from .packages.ordereddict import OrderedDict
+    import collections
+    collections.OrderedDict = OrderedDict
+
+from args import *
 from . import textui
 from . import utils
 from .pipes import piped_in
@@ -19,12 +26,9 @@ from .pipes import piped_in
 
 
 __title__ = 'clint'
-__version__ = '0.2.1'
-__build__ = 0x000201
+__version__ = '0.3.0'
+__build__ = 0x000300
 __author__ = 'Kenneth Reitz'
 __license__ = 'ISC'
-__copyright__ = 'Copyright 2011 Kenneth Reitz'
+__copyright__ = 'Copyright 2012 Kenneth Reitz'
 __docformat__ = 'restructuredtext'
-
-
-args = arguments.Args()