+++ clint/textui/__init__.py
@@ -10,4 +10,7 @@ This module provides the text output helper system.
 
 
 from . import colored
-from core import *
+from . import progress
+from . import prompt
+
+from .core import *