+++ clint/textui/colored.py
@@ -14,6 +14,8 @@ from __future__ import absolute_import
 import re
 import sys
 
+PY3 = sys.version_info[0] >= 3
+
 from ..packages import colorama
 
 __all__ = (
@@ -23,10 +25,17 @@ __all__ = (
 )
 
 COLORS = __all__[:-2]
-DISABLE_COLOR = False
 
-if sys.stdout.isatty():
-    colorama.init(autoreset=True)
+if 'get_ipython' in dir():
+    """
+       when ipython is fired lot of variables like _oh, etc are used.
+       There are so many ways to find current python interpreter is ipython.
+       get_ipython is easiest is most appealing for readers to understand.
+    """
+    DISABLE_COLOR = True
+else:
+    DISABLE_COLOR = False
+
 
 
 class ColoredString(object):
@@ -47,27 +56,36 @@ class ColoredString(object):
 
     def __len__(self):
         return len(self.s)
-        
+
     def __repr__(self):
         return "<%s-string: '%s'>" % (self.color, self.s)
-        
-    def __str__(self):
-        return self.__unicode__().encode('utf8')
-        
+
     def __unicode__(self):
-        return self.color_str
-        
+        value = self.color_str
+        if isinstance(value, bytes):
+            return value.decode('utf8')
+        return value
+
+    if PY3:
+        __str__ = __unicode__
+    else:
+        def __str__(self):
+            return unicode(self).encode('utf8')
+
+    def __iter__(self):
+        return iter(self.color_str)
+
     def __add__(self, other):
         return str(self.color_str) + str(other)
-        
+
     def __radd__(self, other):
         return str(other) + str(self.color_str)
-        
+
     def __mul__(self, other):
         return (self.color_str * other)
-        
-    def split(self, x=' '):
-        return map(self._new, self.s.split(x))
+
+    def split(self, sep=None):
+        return [self._new(s) for s in self.s.split(sep)]
 
     def _new(self, s):
         return ColoredString(self.color, s)